+++ python/maggraph/__init__.py
@@ -50,4 +50,4 @@ __all__ = [
     "open_index",
 ]
 
-__version__ = "0.2.5"
+__version__ = "0.3.0"